task: renaming cta2up component files, types
