+++ src/lib-components/Flexible/GridGalleryCards.vue
@@ -3,19 +3,19 @@ import { computed } from 'vue'
 
 import type { PropType } from 'vue'
 import type { GridGalleryItemType } from '@/types/types'
-import type { FlexibleGridGalleryBlock } from '@/types/flexible_types'
+import type { FlexibleGridGallery } from '@/types/flexible_types'
 
 import GridGallery from '@/lib-components/GridGallery.vue'
 import stripMeapFromURI from '@/utils/stripMeapFromURI'
 
 const { block } = defineProps ({
   block: {
-    type: Object as PropType<FlexibleGridGalleryBlock>,
+    type: Object as PropType<FlexibleGridGallery>,
     default: () => {},
   },
 })
 
-function flattenTimeLineStructure(galleryData: FlexibleGridGalleryBlock) {
+function flattenTimeLineStructure(galleryData: FlexibleGridGallery) {
   const flattenedValues = []
   for (const subitem of galleryData.gridGalleryCards) {
     const obj: GridGalleryItemType = {}
