+++ src/lib-components/index.js
@@ -31,7 +31,7 @@ export { default as CardMeta } from './CardMeta.vue'
 export { default as DividerGeneral } from './DividerGeneral.vue'
 export { default as DividerWayFinder } from './DividerWayFinder.vue'
 export { default as FlexibleCallToAction } from './Flexible/CallToAction.vue'
-export { default as FlexibleCtaBlock2Up } from './Flexible/CtaBlock2Up.vue'
+export { default as FlexibleCallToActionTwoUp } from './Flexible/CallToActionTwoUp.vue'
 export { default as FlexibleGridGalleryCards } from './Flexible/GridGalleryCards.vue'
 export { default as FlexibleMediaGalleryNewLightbox } from './Flexible/MediaGallery/NewLightbox.vue'
 export { default as FooterMain } from './FooterMain.vue'
